backend (3)/backend/requirements.txt
fastapi
flask-cors
pymongo
python-dotenv
transformers
sentencepiece
xgboost
scikit-learn
joblib
pandas
numpy
gunicorn
pymongo
python-dotenv
python-multipart
openai
torch
uvicorn
